Merge the commit and rollback functions to master branch

+++ Menu.sh
@@ -8,7 +8,17 @@ else
 	cd ~/CVS
 fi
 
-#Command to open a repo.
+#Will create the directory containing the older versions of projects if History does not exist.
+if test -d ./.History; then
+	echo ""
+else
+	mkdir .History
+	echo ""
+fi
+
+currentDir=null
+
+#Command to open a directory. $1 determines if the current directory is a repo.
 openDir()
 {
 	#Loops as long as a directory has not been selected or exits if there are no directories to open.
@@ -27,6 +37,10 @@ openDir()
 				if test -d $dir; then
 					#Open directory and display contents as well as current directory location.
 					cd $dir
+					#If the directory is a repo the currentRepo var changes.
+					if [[ $1 -eq 1 ]]; then
+						currentRepo=$dir
+					fi
 					pwd
 					echo ""
 					selectDir=1
@@ -43,7 +57,7 @@ openDir()
 		done
 }
 
-#Will create a repo.
+#Will create a directory. $1 Is the variable to determine if the repo's history files get created.
 createDir()
 {
 	createdDir=0
@@ -57,13 +71,17 @@ createDir()
 			echo ""
 		else
 			mkdir $reponame
+			#If directory is a repo following makes a directory to track the history of the repo. 
+			if [[ $1 -eq 1 ]]; then
+				mkdir ~/CVS/.History/$reponame
+			fi
 			createdDir=1
 			echo ""
 		fi
 	done		
 }
 
-#Deletes a repo.
+#Deletes a directory. $1 is the variable to determine if the repo's history files get deleted. 
 deleteDir()
 {
 	deletedir=0
@@ -73,15 +91,28 @@ deleteDir()
 		if [[ "$(ls -d *)" ]]; then
 			echo "Directory list:"
 			ls -d *
-			dir=null
 			echo "Type the name of the directory to delete."
 			read dir
 			#Tests if directory exists.
 			if test -d $dir; then
-				rm -Rf $dir
-				deletedir=1
-				echo "Successfully deleted repository."
-				echo ""
+				echo "Are you sure you want to delete the directory and all of its files(This is irreversible) [y/n]"
+				read del
+
+				#Ensures the user wishes to delete the repository.
+				if [[ del = "y" ]]; then
+					#Deletes the directory along with the files and sub directories within it.
+					rm -Rf $dir
+					#If directory is a repo following deletes history of repo. 
+					if [[ $2 -eq 1 ]]; then
+						rm -Rf ~/CVS/.History/$dir
+					fi
+					deletedir=1
+					echo "Successfully deleted repository."
+					echo ""
+				else
+					deletedir=1
+					echo "Deleting cancelled."
+				fi
 			else
 				echo "Directory not found. Try again." 
 			fi
@@ -108,14 +139,76 @@ createFile()
 			echo "File already exists."
 			echo ""
 		else
-			#Create file
+			#Create file and exit loop.
 			touch $file
 			fileCreated=1
 		fi
 	done
 }
 
-#REQUIRES LOG IMPLEMENTATION
+#Will commit changes.
+commit()
+{
+	#Adds information of commit to log file.
+	echo "Task:	Commit Files." >> log
+	echo "User:	$USER" >> log
+	echo "Time:	$(date)" >> log
+	echo "Repo:	$(pwd)" >> log
+
+	#if repository is empty print error otherwise continue with commiting files.
+	if [[ $(ls) ]]; then
+		#Opens the .History directory to make a directory with the time/date within the relevant repo.
+		cd ~/CVS/.History/$currentRepo
+		#Stores the date/time in a variable, then makes a directory with the date/time as its name.
+		time=$(date +%Y%m%d%H%M%S)
+		mkdir "$time"
+
+		#Return to the current working directory.
+		cd ~/CVS/$currentRepo
+		#Copy Repo to current version folder.
+		cp -r -f -t ~/CVS/.History/$currentRepo/$time $(pwd)
+
+		echo "Commit Summary:" >> log
+		echo "" >> log
+
+		latestDir=0
+		secondLatestDir=0
+		#read each entry and compare it to the largest value.
+		for entry in ~/CVS/.History/$currentRepo
+		do
+			#If the entry is greater than the newest directory.
+			if [[ $entry -gt latestDir ]]; then
+				secondLatestDir=$latestDir
+				latestDir=$entry
+			fi
+		done
+
+		#If there exists a folder for the second latest directory.
+		if [[ -d ~/CVS/.History/$currentRepo/$secondLatestDir ]]; then
+			#For each entry in the second latest directory in the history directory for the current repo.
+			for entry in ~/CVS/.History/$currentRepo/$secondLatestDir
+			do
+				#If that entry exists in the newly committed directory.
+				if [[ -e ~/CVS/.History/$currentRepo/$latestDir/$entry ]]; then
+					#Append changes between files to the log file.
+					echo "Changes to $entry :" >> log
+					diff $entry ~/CVS/.History/$currentRepo/$latestDir/$entry >> log
+				else
+					echo "File $entry has been deleted." >> log
+				fi			
+			done
+		else
+			#For each entry append the files added to log.
+			for entry in ~/CVS/.History/$currentRepo/$latestDir
+			do
+				echo "Added $entry" >> log
+			done
+		fi
+	else
+		echo "Unable to commit files due to empty repository."
+	fi
+}
+
 #Will open a text file in sublime text.
 openFile()
 {
@@ -144,21 +237,47 @@ openFile()
 	done
 }
 
-#Will append information to a log file when a text file is opened.
+#Will append information to a log file when a text file is opened. $1 is the file name.
 logFileEdit()
 {
-	echo "Task:			Open text file for possible edits." >> log.txt
-	echo "User:			$USER" >> log.txt
-	echo "Time:			$(date)" >> log.txt
-	echo "File:			$1" >> log.txt
-	echo "Directory:	$(pwd)" >> log.txt
+	echo "Task:			Open text file for possible edits." >> log
+	echo "User:			$USER" >> log
+	echo "Time:			$(date)" >> log
+	echo "File:			$1" >> log
+	echo "Directory:	$(pwd)" >> log
 }
 
-#Will commit files
-#commit()
-#{
+#Will allow user to return to an older version of the current repo.
+rollBack()
+{
+	loopExit=0
+	#While the repo has not been rolled back.
+	while [[ loopExit -eq 0 ]]; do
+		#If there are pre existing versions of the current repo.
+		if [[ $(ls ~/CVS/.History/$currentRepo) ]]; then
+			echo "Past versions of $currentRepo (YYYYMMDDHHmmSS):"
+			ls ~/CVS/.History/$currentRepo
+			echo "Enter the name of the version you would like to return to."
+			read ver
+			#If the directory specified by the user exists
+			if [[ -d ~/CVS/.History/$currentRepo/$ver ]]; then
+				#Delete contents of currentRepo.
+				rm -Rf ~/CVS/$currentRepo/*
+				#Copy files from History to current repo.
+				cp -r -f -t ~/CVS/$currentRepo ~/CVS/.History/$currentRepo/$ver
 
-#}
+				echo "Roll back Successfully executed."
+				loopExit=1
+			else
+				echo "Directory does not exist. Try again. "
+			fi
+		else
+			#Exits loop due to inability to roll back.
+			echo "No older versions to return to."
+			loopExit=1
+		fi
+	done
+}
 
 #INCOMPLETE
 #Menu for once a project has been opened.
@@ -169,9 +288,8 @@ projectMenu()
 		echo "Please select a task with relevant number." 
 		echo "1) Create file"
 		echo "2) Open text file"
-		echo "3) Commit file"
-		echo "4) Commit all changes"
-		echo "5) Open directory"
+		echo "3) Commit changes"
+		echo "4) Open directory"
 		echo "0) Exit repository"
 
 		#Handles user input.
@@ -183,9 +301,9 @@ projectMenu()
 			createFile
 		elif [[ option -eq 2 ]]; then
 			openFile
-		#elif [[ option -eq 3 ]]; then
-			
-		elif [[ option -eq 5 ]]; then
+		elif [[ option -eq 3 ]]; then
+			commit
+		elif [[ option -eq 4 ]]; then
 			openDir
 		else
 			echo "Invalid value. Try again."
